+++ qt/qt.go
@@ -9,7 +9,7 @@ import (
 	"github.com/therecipe/qt/quickcontrols2"
 )
 
-func Start() {
+func LaunchApp() {
 	// enable high dpi scaling
 	// useful for devices with high pixel density displays
 	// such as smartphones, retina displays, ...
